-Fixed imports -Moved auth check to NavMenu -Removed unused files
