+++ ClientApp/Auth/Auth.tsx
@@ -43,9 +43,9 @@ export class WebAuthentication implements Auth0Authentication {
 		this.auth0.parseHash((e: Auth0Error, result: Auth0DecodedHash) => {
 			if (result && result.accessToken && result.idToken) {
 				this.setSession(result);
-				history.replace('/home');
+				history.replace('/');
 			} else if (e) {
-				history.replace('/home');
+				history.replace('/');
 				// tslint:disable-next-line:no-console
 				console.error(e);
 				alert(`Error: ${e.error}. Check the console for further details.`);
@@ -61,8 +61,8 @@ export class WebAuthentication implements Auth0Authentication {
 		localStorage.setItem('access_token', accessToken!);
 		localStorage.setItem('id_token', idToken!);
 		localStorage.setItem('expires_at', expiresAt);
-		// navigate to the home route
-		history.replace('/home');
+		// navigate to the base route
+		history.replace('/');
 	}
 
 	@autobind
@@ -71,7 +71,7 @@ export class WebAuthentication implements Auth0Authentication {
 		localStorage.removeItem('access_token');
 		localStorage.removeItem('id_token');
 		localStorage.removeItem('expires_at');
-		// navigate to the home route
-		history.replace('/home');
+		// navigate to the base route
+		history.replace('/');
 	}
 }
