+++ ClientApp/Auth/configuration.tsx
@@ -2,5 +2,5 @@
 export const AUTH_CONFIG: Auth0Config = {
 	domain: 'taproleague.auth0.com',
 	clientId: 'Dy0MGualzLv8MUr3nQH5Q8m7immo0WFp',
-	callbackUrl: 'http://localhost:3000/callback',
+    callbackUrl: 'http://localhost:56856/callback',
 };
