+++ package.json
@@ -4,6 +4,7 @@
   "version": "0.0.0",
   "devDependencies": {
     "@types/auth0-js": "^8.11.6",
+    "@types/autobind-decorator": "^1.3.3",
     "@types/history": "4.6.0",
     "@types/react": "15.0.35",
     "@types/react-dom": "15.5.1",
@@ -11,9 +12,13 @@
     "@types/react-router": "4.0.12",
     "@types/react-router-dom": "4.0.5",
     "@types/webpack-env": "1.13.0",
+    "auth0-js": "^9.8.0",
+    "autobind-decorator": "^2.1.0",
     "aspnet-webpack": "^2.0.1",
     "aspnet-webpack-react": "^3.0.0",
     "awesome-typescript-loader": "3.2.1",
+    "bootstrap": "^4.1.3",
+    "chai": "^4.2.0",
     "css-loader": "0.28.4",
     "event-source-polyfill": "0.0.9",
     "extract-text-webpack-plugin": "2.1.2",
@@ -21,22 +26,16 @@
     "isomorphic-fetch": "2.2.1",
     "jquery": "3.2.1",
     "json-loader": "0.5.4",
+    "popper.js": "^1.14.4",
     "react": "15.6.1",
     "react-dom": "15.6.1",
     "react-hot-loader": "3.0.0-beta.7",
     "react-router-dom": "4.1.1",
+    "shelljs": "^0.8.2",
     "style-loader": "0.18.2",
     "typescript": "2.4.1",
     "url-loader": "^1.1.2",
     "webpack": "2.5.1",
     "webpack-hot-middleware": "2.18.2"
-  },
-  "dependencies": {
-    "auth0-js": "^9.8.0",
-    "autobind-decorator": "^2.1.0",
-    "bootstrap": "^4.1.3",
-    "chai": "^4.2.0",
-    "popper.js": "^1.14.4",
-    "shelljs": "^0.8.2"
   }
 }
